requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.36.0
cryptography==41.0.7
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.5
django-allauth==0.59.0
django-cloudinary-storage==0.3.0
django-environ==0.11.2
django-paypal==2.1
djangorestframework==3.15.2
gunicorn==21.2.0
idna==3.6
oauthlib==3.2.2
packaging==23.2
Pillow==10.0.1
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2025.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
whitenoise==6.6.0
